# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Nesto API
  version: 1.0.0
  description: Hotel Management System API - Backend for Business, Staff, and Customer
    Operations
paths:
...
  /api/v1/bookings/bookings/{booking_ref}/check_in/:
    post:
      operationId: v1_bookings_bookings_check_in_create
      parameters:
      - in: path
        name: booking_ref
        schema:
          type: string
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
  /api/v1/bookings/bookings/{booking_ref}/check_out/:
    post:
      operationId: v1_bookings_bookings_check_out_create
      parameters:
      - in: path
        name: booking_ref
        schema:
          type: string
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookingRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
          description: ''
...
components:
  schemas:
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        booking_number:
          type: integer
          readOnly: true
        booking_id:
          type: string
          readOnly: true
        customer_id:
          type: string
          format: uuid
          readOnly: true
        hotel_name:
          type: string
          maxLength: 255
        room_name:
          type: string
          maxLength: 255
        check_in_date:
          type: string
          format: date
          nullable: true
        check_out_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
        status_display:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - booking_id
      - booking_number
      - created_at
      - customer_id
      - id
      - status_display
      - updated_at
...
    BookingRequest:
      type: object
      properties:
        hotel_name:
          type: string
          maxLength: 255
        room_name:
          type: string
          maxLength: 255
        check_in_date:
          type: string
          format: date
          nullable: true
        check_out_date:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/StatusEnum'
...
    StatusEnum:
      enum:
      - pending
      - confirmed
      - checked_in
      - completed
      - cancelled
      type: string
      description: |-
        * `pending` - Pending
        * `confirmed` - Confirmed
        * `checked_in` - Checked In
        * `completed` - Completed
        * `cancelled` - Cancelled
...
  securitySchemes:
    oauth2:
      type: oauth2
      flows: {}